Clean up code and document more

+++ server.js
@@ -19,6 +19,7 @@ if (process.env.NODE_ENV === 'production') {
   app.get('/', function(req, res) {
     res.sendFile(path.join(__dirname, imgLocation, 'index.html'));
   });
+  // set up for upload
   var storage = multer.diskStorage({
     destination: (req, file, cb) => {
         cb(null, imgLocation + '/images')
@@ -27,7 +28,7 @@ if (process.env.NODE_ENV === 'production') {
         cb(null, String(Date.now()).substring(0, 10) + '-' + file.originalname)
     }
   });
-} else {
+} else { // development version
   imgLocation = 'client/public';
   app.use(express.static(imgLocation));
   var storage = multer.diskStorage({
@@ -55,7 +56,7 @@ const pool = mysql.createPool({
   database: 'leaflet'
 });
 
-// GET route
+// GET route to load images
 app.get('/api/images', (req, res) => {
   pool.query(`select * from ${table}`, (err, rows) => {
     if (err) {
@@ -66,7 +67,7 @@ app.get('/api/images', (req, res) => {
   });
 });
 
-// add image (form data) POST route
+// POST route to store image file (FormData)
 const upload = multer({storage});
 app.post('/api/post', upload.single('image'), (req, res) => {
   if (req.file) {
@@ -77,12 +78,11 @@ app.post('/api/post', upload.single('image'), (req, res) => {
   }
 });
 
-// add image details (JSON) POST route
+// POST route to store image details in SQL database
 app.post('/api/postdetails', (req, res) => {
-  // upload file data into MySQL database
   var sql = `INSERT INTO ${table} (image_title, image_author, src, name, tags) VALUES (?, ?, ?, ?, ?)`;
   let filename = String(Date.now()).substring(0, 10) + '-' + req.body.filename;
-  if (req.body.url) {
+  if (req.body.url) { // upload from URL
     pool.query(sql, [filename, 'admin', req.body.url, req.body.name, JSON.stringify(req.body.tags)], (err, rows) => {
       if (err) {
         console.log("DEBUG: /api/postdetails ERROR MESSAGE: ");
@@ -91,7 +91,7 @@ app.post('/api/postdetails', (req, res) => {
         console.log("DEBUG: /api/postdetails SUCCESS");
       }
     });
-  } else {
+  } else { // upload from device
     pool.query(sql, [filename, 'admin', '/images/'+ filename, req.body.name, JSON.stringify(req.body.tags)], (err, rows) => {
       if (err) {
         console.log("DEBUG: /api/postdetails ERROR MESSAGE: ");
@@ -103,9 +103,9 @@ app.post('/api/postdetails', (req, res) => {
   }
 });
 
-// delete image POST route
+// POST route to delete image details (and file, if it exists)
 app.post('/api/delete', (req, res) => {
-  // upload file data into MySQL database
+  // delete image details from SQL database
   var sql = `delete from ${table} where image_id = ?`;
   console.log(req.body.id);
   console.log(req.body.path);
@@ -116,13 +116,15 @@ app.post('/api/delete', (req, res) => {
     }
   });
   // remove file
-  fs.unlink(imgLocation + '/images/' + req.body.path, (err) => {
-    if (err) {
-      console.log("DEBUG: Unlink file error message");
-      console.error(err);
-      return;
-    } else {
-      console.log("DEBUG: Removed file");
-    }
-  })
+  if (fs.existsSync(imgLocation + '/images/' + req.body.path)) {
+    fs.unlink(imgLocation + '/images/' + req.body.path, (err) => {
+      if (err) {
+        console.log("DEBUG: Unlink file error message");
+        console.error(err);
+        return;
+      } else {
+        console.log("DEBUG: Removed file");
+      }
+    })
+  }
 });
